DONE: Fixed problems at DB
Now we store also store Categories so our user won't have to download
the list everytime (plus it makes sense cause we won't update categories
everyday or something).

Places are now stored in the DB and the thing with the problem of
PrimaryKeys is now solved.

I think Places module is ready to roll (we need more data though)